requirements.txt
# 核心依赖
langchain==0.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
httpx==0.25.2
pydantic-settings==2.1.0
chromadb==0.4.18
python-jose==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23

# 文档处理相关
unstructured>=0.10.30
pymupdf>=1.23.6
python-docx>=0.8.11

# 可视化相关
streamlit>=1.29.0
plotly>=5.18.0
echarts-python>=0.1.3